gcmc example working
