+++ python/dgl/dataloading/__init__.py
@@ -1,4 +1,8 @@
-"""Data loading modules"""
+"""Data loading modules.
+
+NOTE: this module is experimental and the interfaces may be subject to changes in
+future releases.
+"""
 from .neighbor import *
 from .dataloader import *
